+++ src/RBR.hpp
@@ -34,8 +34,7 @@ namespace rbr {
     bool is_rendering_particles();
 
     void change_camera(void* p, uint32_t cameraType);
-    const M4 get_horizon_lock_matrix();
-    const M4& get_seat_translation_matrix();
+    const M4& get_horizon_lock_matrix();
 
     // Hookable functions
     void __fastcall render(void* p);